FIX failing inference
