+++ .github/workflows/lint.yml
@@ -2,7 +2,7 @@
 name: Lint
 
 on:
-  # push:
+  push:
   pull_request:
 
 jobs:
